moving non-chat related components to ui/
