+++ llama_stack/ui/components/chat/chat.tsx
@@ -11,9 +11,9 @@ import { ArrowDown, ThumbsDown, ThumbsUp } from "lucide-react"
 
 import { cn } from "@/lib/utils"
 import { useAutoScroll } from "@/hooks/use-auto-scroll"
-import { Button } from "@/components/chat/button"
+import { Button } from "@/components/ui/button"
 import { type Message } from "@/components/chat/chat-message"
-import { CopyButton } from "@/components/chat/copy-button"
+import { CopyButton } from "@/components/ui/copy-button"
 import { MessageInput } from "@/components/chat/message-input"
 import { MessageList } from "@/components/chat/message-list"
 import { PromptSuggestions } from "@/components/chat/prompt-suggestions"
