+++ llama_stack/ui/components/chat/markdown-renderer.tsx
@@ -3,7 +3,7 @@ import Markdown from "react-markdown"
 import remarkGfm from "remark-gfm"
 
 import { cn } from "@/lib/utils"
-import { CopyButton } from "@/components/chat/copy-button"
+import { CopyButton } from "@/components/ui/copy-button"
 
 interface MarkdownRendererProps {
   children: string
