+++ llama_stack/ui/components/chat/message-input.tsx
@@ -8,9 +8,9 @@ import { omit } from "remeda"
 import { cn } from "@/lib/utils"
 import { useAudioRecording } from "@/hooks/use-audio-recording"
 import { useAutosizeTextArea } from "@/hooks/use-autosize-textarea"
-import { AudioVisualizer } from "@/components/chat/audio-visualizer"
-import { Button } from "@/components/chat/button"
-import { FilePreview } from "@/components/chat/file-preview"
+import { AudioVisualizer } from "@/components/ui/audio-visualizer"
+import { Button } from "@/components/ui/button"
+import { FilePreview } from "@/components/ui/file-preview"
 import { InterruptPrompt } from "@/components/chat/interrupt-prompt"
 
 interface MessageInputBaseProps
