+++ llama_stack/ui/components/ui/collapsible.tsx
