+++ llama_stack/ui/components/ui/copy-button.tsx
@@ -4,7 +4,7 @@ import { Check, Copy } from "lucide-react"
 
 import { cn } from "@/lib/utils"
 import { useCopyToClipboard } from "@/hooks/use-copy-to-clipboard"
-import { Button } from "@/components/chat/button"
+import { Button } from "@/components/ui/button"
 
 type CopyButtonProps = {
   content: string
