+++ llama_stack/ui/components/ui/file-preview.tsx
